+++ flutter_theoplayer_sdk/lib/theoplayer.dart
@@ -629,6 +629,7 @@ class _PlayerPlatformActivityServiceListener implements PlatformActivityServiceL
 
     if (!player.allowAutomaticPictureInPicture()) {
       debugLog("THEOplayer_$playerID: PlayerPlatformActivityServiceListener onExitPictureInPicture not for me");
+      return;
     }
 
     player._setPresentationMode(player._presentationModeBeforePip);
@@ -643,6 +644,7 @@ class _PlayerPlatformActivityServiceListener implements PlatformActivityServiceL
 
     if (!player.allowAutomaticPictureInPicture()) {
       debugLog("THEOplayer_$playerID: PlayerPlatformActivityServiceListener onUserLeaveHint not for me");
+      return;
     }
 
     player._setPresentationMode(PresentationMode.PIP, userTriggered: false);
@@ -660,6 +662,34 @@ class _PlayerPlatformActivityServiceListener implements PlatformActivityServiceL
 
       Navigator.of(player._currentContext, rootNavigator: true).push(pipRoute);
     }
+  }
+
+  @override
+  void onPlayActionReceived() {
+    debugLog("THEOplayer_$playerID:: PlayerPlatformActivityServiceListener onPlayActionReceived");
+
+
+    if (!player.allowAutomaticPictureInPicture()) {
+      debugLog("THEOplayer_$playerID: PlayerPlatformActivityServiceListener onPlayActionReceived not for me");
+      return;
+    }
+
+    player.play();
+
+  }
+
+  @override
+  void onPauseActionReceived() {
+    debugLog("THEOplayer_$playerID:: PlayerPlatformActivityServiceListener onPauseActionReceived");
+
+
+    if (!player.allowAutomaticPictureInPicture()) {
+      debugLog("THEOplayer_$playerID: PlayerPlatformActivityServiceListener onPauseActionReceived not for me");
+      return;
+    }
+
+    player.pause();
+
   }
 
   @override